fix some tests and cleanup
