+++ src/interface/src/app/plan/plan-summary/saved-scenarios/saved-scenarios.component.html
@@ -7,6 +7,7 @@
           matSuffix
           mat-raised-button
           color="primary"
+          data-id="new-scenario"
           (click)="openConfig()">
           <mat-icon class="material-symbols-outlined">add_box</mat-icon>
           NEW SCENARIO
